added m_allocator member to basic_json and parser class

+++ test/src/unit.cpp
@@ -966,7 +966,7 @@ TEST_CASE("constructors")
     SECTION("create an array of n copies of a given value")
     {
         json v = {1, "foo", 34.23, {1, 2, 3}, {{"A", 1}, {"B", 2u}}};
-        json arr(3, v);
+        json arr(static_cast<size_t>(3), v);
         CHECK(arr.size() == 3);
         for (auto& x : arr)
         {
